+++ src/index.d.ts
@@ -47,8 +47,7 @@ export interface IFlagship {
   config: FlagshipSdkConfig;
     log: any;
     envId: string;
-    newVisitor(id: string, context: FlagshipVisitorContext): IFlagshipVisitor; // deprecated
-    createVisitor(id: string, context: FlagshipVisitorContext): IFlagshipVisitor;
+    newVisitor(id: string, context: FlagshipVisitorContext): IFlagshipVisitor;
 }
 
 export interface FlagshipNodeSdk {