+++ RELEASENOTES.md
@@ -6,6 +6,12 @@
 
 - Add [`getModificationInfo`](README.md#getModificationInfo) function.
 
+### Breaking changes ⚠️
+
+- Function removed:
+
+  - createVisitor --> [use `newVisitor` instead](README.md#newVisitor)
+
 ## ➡️ Version 1.0.1
 
 ### Bug fixes 🐛